hal: Rework WDC-based timer driver
The previous design assumed that the cascaded counters worked in a
typical and sensible way.

This was not the case.

For the PSoC4100S and PSoC4100S Plus, the WDC counters cascade not on
rolling over, but on match. This is true no matter if they are in
clear-on-match or free-run mode. This means that if we want to use the
counters together as a 64-bit counter, we cannot use the interrupts for
arbitrary timestamps.

The new design uses the 32-bit free-run counter coupled with an
AtomicU32 as the 64-bit timestamp source. That leaves the two 16-bit
counters to act as alarms. Because all WDC counters run off the same
clock, they will stay in lock-step with the 32-bit counter used to keep
track of the timestamp.

Both 16-bit counters are set up in free-run mode so they stay "in phase"
with the 32-bit counter. The 16-bit counters are set to trigger on the
difference between "now" and the timestamp at the point where the alarm
was started. Any alarm set more than 1.6 seconds into the future will
trigger a few times in-between, but that is fine. There will be some
additionally delay added to the required delay due to setup time. This
setup time is dominated by required delays for resetting and
enabling/disabling the timers. These delays are minimized by starting
these operations as early as feasibly possible; when an alarm is being
disabled.