Disable clocks that require external pins
Currently we have no way of passing in a pin during initialization as to
get the pin peripheral you need to perform initialization, resulting in
a chicken and egg scenario. We might as well remove the fields as they
won't be public anyway when we do support binding a pin for these
features.